requirements.txt
cbor2==5.4.6
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
docker==6.0.1
idna==3.4
ifaddr==0.2.0
luma==2.1.5
luma.core==2.4.0
luma.oled==3.12.0
packaging==23.1
Pillow==9.5.0
pycparser==2.18
pyftdi==0.54.0
pyparsing==2.2.0
PyQRCode==1.2.1
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
requests==2.29.0
roonapi==0.1.4
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.4.2
spidev==3.6
urllib3==1.26.15
websocket-client==1.5.1
